Content: 2 new story locations The Huaca and The Panopticon (by yapludepatte and Virrno)
Content: 2 new story locations The Huaca and The Panopticon (by yapludepatte and Virrno)
This most likely need server admins and players to allow those structures to generate at their locations if those chunks are already generated prior to the update.
Also regenerate the crossroads to allow people to find the map to those new structures.